+++ week4_login/src/App.jsx
@@ -1,17 +1,23 @@
 import { ThemeProvider } from '@emotion/react';
 import './App.css';
-import * as React from 'react';
+import { React, useState } from 'react';
 import Login from './pages/Login';
 import theme from './styles/theme';
 import GlobalStyle from './styles/GlobalStyle';
-
+import Join from './pages/Join';
 import Router from '../Router';
+import Main from './pages/Main';
+import MyPage from './pages/MyPage';
+
 function App() {
   return (
     <ThemeProvider theme={theme}>
       <GlobalStyle />
-      <Login />
       <Router>
+        <Main />
+        <Login />
+        <MyPage />
+        <Join />
       </Router>
     </ThemeProvider>
   );
