+++ week4_login/src/components/common/Container.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import styled from '@emotion/styled';
+
+const Container = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+`;
+export default Container;
